Closes: #6223, #6106, #6281.
Offer the user the ability to graphically choose a new resolution from
an ordered list, by clicking an icon in the panel. This user-selected
value is then stored on disk, to be re-used next time X restarts.

Change Vmware screen resizing behavior: the default 800x600 is still
considered a bit ridiculous, but we now force it to 1024x768, instead
of using what xrandr reports as the maximum available size (the user
can use the tool described above to change it some more). Also, _do_
check for 800x600 before blindly resizing.

Set a reasonable DPI value if the current one is grossly wrong (this
does not seem to work on VMWare, but the issue we're fixing doesn't
show up on VMWare anyway).


git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/pkgs@23022 8ebf8b4e-06f3-0310-961f-89f74618be54