I don't see this on my side, but chmod'ing /var/log/FuzzyOcr.log to 644 certainly cannot hurt anything (Closes: #7489).
Juan, for the sake of my better understanding the problem, could you
tell me what "ls -l /var/log/FuzzyOcr.log" was returning *before* you
chmod'ed it ? The postinst for untangle-fuzzyocr sets the owner as
spamc, and it's pretty unusual for a file to not be +w for its owner,
so I'm wondering what was going on... If this is something else
altogether, I'll re-open accordingly.

Thanks a lot for your time.


git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/pkgs@25791 8ebf8b4e-06f3-0310-961f-89f74618be54