Allow for ut-qemu-run to handle non-NGFW related setups
Turns out the new QEMU framework is so popular that developers are using
it for VMs not related at all to NGFW work, so we can't insist on having
-c or -b present on the command-line.

Instead we settle on requiring both -c and -b when -u is
present (meaning "we are a NGFW server"). For a client-type setup, the
caller is free to pass -c (to connect to a NGFW bridge) or not (for
running any type of VM).