wireless: Add client wireless support
If a wireless interface is configured as a wireless client, write out a
wpa_supplicant configuration file and then reference that file from
the interface's entry in /etc/network/interfaces so that wpa_supplicant
will be started/stopped when the interface is brought up/down.

If the client wireless interface is configured as an addressed (maybe WAN)
interface, add the wpa-conf line to its normal /etc/network/interfaces
entry.

If the client wireless interface is configured as a bridged interface, we
need to create the bridged interface entry in /etc/network/interfaces and
add the wpa-conf line to that.  wpa_supplicant needs the interface to be
brought up in order to work.

Also make untangle-sync-settings depend on the wpasupplicant package.

NGFW-5922