wireless: Use if-pre-up.d/if-post-down.d hooks to manage hostapd
If the wireless interface is configured, write out the hostapd
configuration file and then reference that file from the interface's
entry in /etc/network/interfaces.  That way hostapd will be
started/stopped automatically when the interface is brought up/down.

If the wireless interface is configured as an addressed interface, add
the hostapd line to its normal /etc/network/interfaces entry.

If the wireless interface is configued as a bridged interface, create
the bridged interface entry in /etc/network/interfaces and add
the hostapd line to that.