01minwrites: Explicitly remount as rw
When booting linksys firmware, the kernel seems to initially mount
the rootfs as ro.  But fairly early in userspace 01minwrites
calls 'mount -o remount,relatime,nodiratime' which on a jessie build
seems to remount as rw even though it doesn't explicitly say to do so.
After upgrading a linksys build to stretch, this call to mount doesn't
automatically remount as rw.  I had to add rw to the mount options in
order to get a writeable filesystem.  I cannot explain the change in
behavior.