kiosk: Wait until the uvm is running before starting x
systemd is starting x and uvm in parallel.  If x wins
the race it'll start chromium before the uvm is running
and the user will see "server not found".

Wait to start x until we know the uvm is running.  Then
chromium will be able to connect for sure.

NGFW-11475 #code-complete