Add initial route_manager
route_manager currently manages the following files:

/etc/iproute2/rt_tables
This file must contain any custom route tables we plan to use.
route_manager adds wan.[interfaceId] tables and a balance table.

/etc/config/[ifup.d|ifdown.d]/10-default-route
These scripts wait for ifup and ifdown events sent by netifd.  If the
events are for wan interfaces, it makes sure that a priority 1000000
default route to one of the available "up" wan ports always exists.

/etc/config/[ifup.d|ifdown.d]/20-wan-balancer
These scripts wait for ifup and ifdown events sent by netifd.  If the
events are for wan interfaces, it rebuilds the balance table to ensure it
always includes all of the "up" wan ports.

MFW-106