network_manager: Set force_link to 0 for static addresses
Don't force the link for interface's with static addresses.  This
means that netifd will send ifup/ifdown alerts when the cable is
plugged/unplugged.  This will allow wan balancer to know to
rewrite the balance table on cable plugs/unplugs.

The somewhat unfortunate side effect of this is that if the
cable is unplugged, the interface won't be configured.  So if you
run 'ifconfig eth0' while eth0's cable is unplugged, it won't have
its static address applied.

MFW-106