board_util: Set the mac address for switched interfaces on linksys wrt3200
The switched interfaces on linksys boxes come up with a random mac address
every boot unless we set it.  This is because the device the switch is
connected to, eth1, doesn't have a mac address at the time that the dsa
switch driver creates the switched interface.

Set the switched interface mac addresses to be the same as eth1's.  This
seems a bit odd, but thats how the Turris Omnia works out of the box.  We
can always change the macs we set latter.

MFW-104