openwrt: wireless_manager: Enumerate wireless interfaces via phy
The wireless phys (phy0, phy1, etc) represent the actual wireless hardware
of the device, while the wireless interfaces (wlan0, wlan1, etc) are merely
a construct of the kernel.  As such, when a phy is disabled, its
corresponding interface may be removed.  This was causing us a problem as
our code was trying to determine the phy for a given interface after the
phy had been disabled.  The wireless interface didn't exist and we didn't
know how to write an entry for the radio in /etc/config/wireless.

When enumerating wireless interfaces in the system (during
'sync-settings -c' for example) we need to discover the wireless phys in
the system and then enumerate the wireless interfaces that correspond to
them.  By ensuring wireless interfaces are named with the same index as
the corresponding phy (phy0 -> wlan0, phy1 -> wlan1, etc) we make phy
to/from interface lookups into easy string replaces.  We can now
successfully create radio entries in /etc/config/wireless regardless of
whether the phys are currently enabled.

MFW-156