wireless_manager: Tweak for radios that support 2.4 and 5 Ghz
Before this commit we assumed a radio that supported 5 GHz ONLY
supported 5 GHz, which is a bad assumption.  Some radios support
2.4 and 5 GHz.  Now we check if the user supplied channel is valid
for the particular card and write /etc/config/wireless correctly
no matter the band.

Ideally we would only allow the customer to select channels
appropriate for the card they are configuring, but that'll have
to be an exercise for another day.