network_manager: Support multiple default external/wan interfaces
On some devices we may want to define more than one interface as a default
wan.  Now we can define a list of interfaces to configure as wans instead
of just a single interface.  We also have the ability to define a list
of interfaces that should be bridged together for the internal network.
The first interface in the list will be the addressed interface and all of
the others will be bridged to it.  Any wlanX interfaces will also be
bridged to internal by default.

On the espressobin, we now define two default wan interfaces named WAN0
and WAN1.