logs: Send json in log prefix
The json will contain the table, chain, ruleid, and action.
In the case of a WAN_POLICY action, the policy is also included.

If we make a routing decision based on the cache, we still send
a rule log, but the ruleId is -1.

If a session is not routed via any policy and falls back to the
system default wan, the ruleId and policy are -2.