Update interface_manager.py for Citrix
We need to restore both interface marks on re-injected TCP packets.
Our reinject-restore-marks chain will extract the interface marks
from the source MAC address where the UVM stores them and set them
in the packet mark. This requires our modified xt_mac.ko matcher
module that has been enhanced to allow matching on the individual
bytes of the MAC address. When the first byte of the comparison
MAC address is 0xFF the module will use the second byte of the
address as the index to be compared. So bytes at index 0 and 1 are
used to trigger our special match logic. 2 and 3 are unused.
4 = dst interface and 5 = src interface.