route_manager: Report rule id for traffic routed via cache
Now, when traffic is routed via a rule, we add the src/dst addresses
to a set for that particular rule.  Then, when traffic is routed by
a cache lookup, we can check so see if that src/dst combo is in a
particular rule table, and if it is, send an event to packetd to
update the rule id for that session.