Fix GeoIP database update cron job
The company (MaxMind) that provides the GeoIP database we use has
recently changed the availability of the free database so an account
is required to download the file. I modified our update cron job to now
download the file from an Untangle server. I also added some additional
error detection and handling in the script, and used a wait in places
where the script will do the wrong thing if run from an interactive
shell where child procs run in a sub-shell that doesn't block the
main script.

NGFW-12778 #review