route_manager: Only route new connections
Theoretically, only new connections should make it to this point
as established connections should be caught by the destination mark
check above it.  But there are some rare cases where a non-new
connection might make it through.  Usually, this is the invalid case.

So, only route new connections, but log other ct states so we can have
an idea if something is amiss.  A handful of invalid packets isn't
anything to worry about.  But established and related connections making
it throught the destination mark check could indicate a problem.