vendor1: some geoip tweaks
There is only one implementation for writing MMDB files:
https://metacpan.org/release/MaxMind-DB-Writer, from
https://github.com/maxmind/MaxMind-DB-Writer-perl. However, the basic
example described at
https://blog.maxmind.com/2015/09/29/building-your-own-mmdb-database-for-fun-and-profit/
is non functional. No matter what I try to write to the DB, it always
results in a "PerlIO_printf operation failed: Unknown error -1"
exception. I've spent a couple of hours troubleshooting this, to no
avail.

Also, the writer isn't packaged for Debian, so it needs to be installed
from source via CPAN, and since it contains C code and depend on
libraries that contain C++ code, we need to install all the compilers
from build-essential, which seems a bit unproductive as the end goal
here seems to be about saving an extra 60M of disk space.

Comment: NGFW-13137