u-clamav-config: Include default buster clamd.conf
Users that have updated through multiple untangle versions have
a stale clamd.conf.  On upgrade, install a buster based clamd.conf
file to get the whole install base up to date.  We will only pull in
the cached buster config file if we are updating to 16.0, so that on
new installs that may include a newer clamav-daemon package, we will
allow that package to install its updated clamd.conf

NGFW-13068