wireless_manager: Handle undefined wirelessVisibility
On upgrade to 15.1 we run sync-settings as part of the sync-settings
package postinst.  Customers with wireless interfaces that have not
made a network settings change post-14.2 will not have the
wirelessVisibility field defined in their network.js, which means
that the postinst sync-settings call will fail, which can cause
the upgrade to 15.1 to stall.  If wirelessVisibility isn't defined
for a wireless interface, just write out a 0.

NGFW-13205