Ensure depmod finishes before the uvm starts
One of the things that untangle-hardware-config does is call
depmod -a.  On most boots this takes almost no time at all, but
on the boot after we've changed kernels, this depmod call might
take a while.  And the module loading that we do prior to starting
the uvm depends on depmod being finished.  So explicitly ensure that
the untangle-linux-config service finishes (ensuring that depmod
has finished) before starting the untangle-vm service.

NGFW-13231