route_manager: Improve and move enabled_wan to network_util
When checking if something is an enabled wan, we also need to
check if its addressed or not.  If it isn't addressed, then its
not a wan, even if the wan field is set to true.  Also, move this
function to network_util so other managers can use it.

MFW-1137