Add flag to control application of default network settings
Added a flag file to control when we copy and sync the default
network settings. This will ensure our defaults get applied
if this package is installed after the uvm is started for the
first time which creates the default network.js which otherwise
would prevent the default settings from being applied.