u-development-runtime: prevent removal of python-pycurl and python-pil on buster
python3-pycurl and python3-pil are already listed as dependencies for
u-development-runtime, so apt's resolver will see them as uneeded in
"python-foo | python3-foo".

Since python3.9 is only in bullseye but not in buster, we choose to base
the alternative dependencies on that instead: this will ensure that no
python-foo is installed on bullseye.